Refactor: 변수 선언 모듈화
