+++ js/declaration.js
@@ -1,0 +1,5 @@
+export const emailDiv = document.querySelector('#email');
+export const emailInput = document.querySelector('.email-input');
+export const pwdInput = document.querySelector('.password-input');
+export const signinBtn = document.querySelector('#signinBtn');
+export const pwdEyeIcon = document.querySelector('.PwdEyeIcon');
